fix(playground): 在vue演练场中添加 mateChat示例
