+++ package.json
@@ -12,6 +12,7 @@
     "build:remoter": "pnpm -F @opentiny/next-remoter build",
     "build:site": "pnpm -F @opentiny/next-remoter build:site",
     "dev:remoter": "pnpm -F @opentiny/next-remoter dev",
+    "dev:vue-playground": "pnpm -F vue-playground dev",
     "pub:next-sdk": "pnpm -F @opentiny/next-sdk publish --no-git-checks",
     "pub:remoter": "pnpm -F @opentiny/next-remoter publish --no-git-checks"
   },
