+++ packages/vue-playground/README.md
@@ -1,0 +1,3 @@
+# Chat UI 的演练场
+
+演示`"@opentiny/next-remoter"` 库导出的 `useNextAgent` 如何适配在其它UI上。目前已适配 'matechat' | 'antx' | 'elementplusx' | 'tdchat' 。
