+++ packages/vue-playground/index.html
@@ -1,12 +1,13 @@
 <!doctype html>
-<html lang="zh-CN">
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>3rd Remoter</title>
+    <title>vue-playground</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./main.ts"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
